Change implementation of in_progress to flag
For future implementation of a fail-safe mechanism an additional flag is
needed. Since 'in_progress' only needs one bit, the meaning of the
internal 'in_progress' variable is changed to a flags register and
'in_progress' is represented by the flag setting the LSB.
This way 7 additional flags can be implemented without the need for
additional variables.

Signed-off-by: Andreas Reichel <andreas.reichel.ext@siemens.com>